backend/requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Authentication
PyJWT==2.8.0
cryptography>=3.4.8

# HTTP requests
requests>=2.28.0
httpx>=0.24.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Image processing
Pillow>=9.0.0

# Environment and configuration
python-dotenv>=0.19.0

# Data validation
marshmallow>=3.19.0

# Utilities
python-dateutil>=2.8.0

# Development and testing
pytest>=7.0.0
pytest-flask>=1.2.0
black>=22.0.0
flake8>=5.0.0

# Production server
gunicorn>=20.1.0

# Caching
redis>=4.0.0

# Logging
structlog>=22.0.0
